refactor: replace dto errors
