+++ src/main.ts
@@ -3,7 +3,9 @@ import { AppModule } from './app.module';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
-  const PORT = process.env.PORT || 4000;
+  const PORT = process.env.PORT ?? 5000;
+  console.log(PORT)
   await app.listen(PORT);
+
 }
 bootstrap();
