+++ src/user/user.service.ts
@@ -1,8 +1,8 @@
 import { ForbiddenException, Injectable } from '@nestjs/common';
 import { User } from './user.entity';
 import { v4 as uuidv4 } from 'uuid';
-import { CreateUserDto } from './create-user.dto';
-import { UpdatePasswordDto } from './update-user.dto';
+import { CreateUserDto } from './dto/create-user.dto';
+import { UpdatePasswordDto } from './dto/update-user.dto';
 
 @Injectable()
 export class UserService {
